refactor: rename automator to roger and remove the oclif roger
