--- packages/automator/.gitignore
@@ -1,4 +1,0 @@
-/artifacts/
-/browser/
-/out/
-/progs/
