--- packages/automator/package.json
@@ -1,49 +1,0 @@
-{
-  "name": "@penrose/automator",
-  "version": "3.0.0-alpha.4",
-  "description": "",
-  "type": "module",
-  "main": "index.tsx",
-  "scripts": {
-    "build": "tsc",
-    "build-decls": ":",
-    "start": "tsx ./index.ts",
-    "generate-site": "yarn clean && yarn start batch registry.json artifacts --render=browser --src-prefix=progs --folders --cross-energy",
-    "clean": "rimraf artifacts browser",
-    "typecheck": "tsc --noEmit"
-  },
-  "bin": {
-    "automator": "./bin/run.js"
-  },
-  "nx": {
-    "targets": {
-      "build": {
-        "dependsOn": [
-          "^build",
-          "^build-decls"
-        ]
-      }
-    }
-  },
-  "author": "Penrose Team (https://penrose.cs.cmu.edu/)",
-  "license": "MIT",
-  "dependencies": {
-    "@happy-dom/global-registrator": "^9.18.3",
-    "@penrose/core": "^3.0.0-alpha.0",
-    "canvas": "^2.11.2",
-    "chalk": "^3.0.0",
-    "child_process": "^1.0.2",
-    "convert-hrtime": "^3.0.0",
-    "global-jsdom": "^8.8.0",
-    "jsdom": "^22.0.0",
-    "neodoc": "^2.0.2",
-    "prettier": "2.2.1",
-    "regenerator-runtime": "^0.13.11",
-    "true-myth": "^4.1.1",
-    "tsx": "^3.12.7",
-    "uniqid": "^5.4.0"
-  },
-  "devDependencies": {
-    "@types/node": "^12.12.68"
-  }
-}
