--- packages/automator/tsconfig.json
@@ -1,28 +1,0 @@
-{
-  "compilerOptions": {
-    "module": "es2022",
-    "allowJs": true,
-    "esModuleInterop": true,
-    "target": "es2017",
-    "moduleResolution": "Node",
-    "sourceMap": true,
-    "outDir": "dist",
-    "jsx": "react",
-    "typeRoots": ["./node_modules/@types"],
-    "lib": ["es5", "es6", "es7", "esnext", "dom"],
-    "forceConsistentCasingInFileNames": true,
-    "importHelpers": true,
-    "suppressImplicitAnyIndexErrors": true,
-    "allowSyntheticDefaultImports": true,
-    "resolveJsonModule": true,
-    "noUnusedLocals": false,
-    "skipLibCheck": true,
-    "downlevelIteration": true,
-    "types": ["jest", "node", "tslib"],
-    "isolatedModules": true,
-    "declaration": true,
-    "declarationMap": true,
-    "noImplicitThis": true,
-    "strictNullChecks": true
-  }
-}
