+++ packages/docs-site/package.json
@@ -5,7 +5,7 @@
   "scripts": {
     "build": "vitepress build",
     "build-decls": ":",
-    "build-shapedefs": "shx mkdir -p ./public/try/ && shx cp -r ../editor/dist/* ./public/try && shx rm -f ./src/shapedefs.json && cd ../automator && yarn start shapedefs ../docs-site/src/shapedefs.json",
+    "build-shapedefs": "shx mkdir -p ./public/try/ && shx cp -r ../editor/dist/* ./public/try && shx rm -f ./src/shapedefs.json && cd ../roger && yarn start shapedefs ../docs-site/src/shapedefs.json",
     "dev": "vitepress dev",
     "preview": "vitepress preview",
     "typecheck": ":"
@@ -42,7 +42,7 @@
     "veaury": "^2.3.11"
   },
   "devDependencies": {
-    "@penrose/automator": "^3.0.0-alpha.1",
+    "@penrose/roger": "^3.0.0-alpha.1",
     "@penrose/editor": "^3.0.0-alpha.0",
     "markdown-it-katex": "^2.0.3",
     "shx": "^0.3.3",
