+++ packages/roger/.gitignore
@@ -1,9 +1,4 @@
-*-debug.log
-*-error.log
-/.nyc_output
-/dist
-/lib
-/package-lock.json
-/tmp
-node_modules
-oclif.manifest.json
+/artifacts/
+/browser/
+/out/
+/progs/
