--- packages/roger/bin/run
@@ -1,5 +1,0 @@
-#!/usr/bin/env node
-
-const oclif = require('@oclif/core')
-
-oclif.run().then(require('@oclif/core/flush')).catch(require('@oclif/core/handle'))
