+++ packages/roger/docs/penrose-artifacts.png
