+++ packages/roger/package.json
@@ -1,68 +1,50 @@
 {
   "name": "@penrose/roger",
-  "version": "3.0.0-alpha.0",
-  "description": "The Penrose CLI",
-  "author": "Penrose Team (https://penrose.cs.cmu.edu/)",
-  "bin": {
-    "roger": "./bin/run"
+  "version": "3.0.0-alpha.4",
+  "description": "",
+  "type": "module",
+  "main": "index.tsx",
+  "scripts": {
+    "build": "tsc",
+    "build-decls": ":",
+    "start": "tsx ./index.ts",
+    "generate-site": "yarn clean && yarn start batch registry.json artifacts --render=browser --src-prefix=progs --folders --cross-energy",
+    "clean": "rimraf artifacts browser",
+    "typecheck": "tsc --noEmit"
   },
-  "homepage": "https://github.com/penrose/penrose",
+  "bin": {
+    "roger": "./bin/run.js"
+  },
+  "nx": {
+    "targets": {
+      "build": {
+        "dependsOn": [
+          "^build",
+          "^build-decls"
+        ]
+      }
+    }
+  },
+  "author": "Penrose Team (https://penrose.cs.cmu.edu/)",
   "license": "MIT",
-  "main": "dist/index.js",
-  "repository": "penrose/penrose",
-  "files": [
-    "bin",
-    "dist",
-    "oclif.manifest.json"
-  ],
   "dependencies": {
-    "@oclif/core": "^1",
-    "@oclif/plugin-help": "^5",
-    "@oclif/plugin-plugins": "^2.0.1",
+    "canvas": "^2.8.0",
+    "chalk": "^3.0.0",
+    "child_process": "^1.0.2",
+    "convert-hrtime": "^3.0.0",
+    "global-jsdom": "^8.8.0",
+    "jsdom": "^22.0.0",
+    "neodoc": "^2.0.2",
+    "prettier": "2.2.1",
+    "regenerator-runtime": "^0.13.11",
+    "true-myth": "^4.1.1",
+    "tsx": "^3.12.7",
+    "uniqid": "^5.4.0",
     "chokidar": "^3.5.3",
     "ws": "^8.6.0"
   },
   "devDependencies": {
-    "@oclif/test": "^2",
-    "@types/chai": "^4",
-    "@types/mocha": "^9.0.0",
-    "@types/node": "^16.9.4",
-    "@types/ws": "^8.5.3",
-    "chai": "^4",
-    "eslint": "^7.32.0",
-    "eslint-config-oclif": "^4",
-    "eslint-config-oclif-typescript": "^1.0.2",
-    "globby": "^11",
-    "mocha": "^9",
-    "oclif": "^3",
-    "shx": "^0.3.3",
-    "ts-node": "^10.2.1",
-    "tslib": "^2.3.1"
-  },
-  "oclif": {
-    "bin": "roger",
-    "dirname": "roger",
-    "commands": "./dist/commands",
-    "plugins": [],
-    "topicSeparator": " ",
-    "topics": {}
-  },
-  "scripts": {
-    "build": "shx rm -rf dist && tsc -b",
-    "build-decls": ":",
-    "typecheck": ":",
-    "postpack": "shx rm -f oclif.manifest.json",
-    "prepack": "yarn build && oclif manifest && oclif readme && prettier --write README.md",
-    "version": "oclif readme && git add README.md",
-    "preinstall-global": "yarn build",
-    "install-global": "yarn link"
-  },
-  "engines": {
-    "node": ">=12.0.0"
-  },
-  "bugs": "https://github.com/penrose/penrose/issues",
-  "keywords": [
-    "oclif"
-  ],
-  "types": "dist/index.d.ts"
+    "@types/ws": "^8.5.4",
+    "@types/node": "^12.12.68"
+  }
 }
