+++ packages/roger/trio.json
